Update MWSClient
For one of my clients I had to integrate the Dutch market place, and I realized that the class did not support it. So I had to manually insert it into the source, which is an unclean operation.
So I thought of integrating the possibility of adding the marketplace in the manufacturer.
Also, the method was missing from the class to check the status of the service. So I added the getServiceStatus method.
newLine from 1272 to 1286